Animation work and equip handler
The shooting handler turned into the equip handler and does stuff like checking the inventory and giving us values to work with elsewhere.
