--- Assets/Art/Animations/Humanoids/Player/Arm Masks/Pistol Idle.anim.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7c0e0de6e1cf2ca4bbace63a26c56db0
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 7400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
