+++ Assets/Art/Animations/Humanoids/Player/Arm Masks/Pistol.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: bbe3165524bc50f40a94d85fbc67524e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
