+++ Assets/Art/Animations/Humanoids/Player/Arm Masks/Pistol/Pistol Aim.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ce1259ee86357ec428493401280a826d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
