+++ Assets/Art/Animations/Humanoids/Player/Arm Masks/Pistol/Pistol Shoot.anim.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 85ba82a5c5fa45c448f4849edad72513
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 7400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
