--- Assets/Scriptable Objects/Flashlight.asset
@@ -1,29 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2fd88fe6ee8f524418b24e29f14bd1bb, type: 3}
-  m_Name: Flashlight
-  m_EditorClassIdentifier: 
-  _index: 5
-  itemName: Flashlight
-  description: 
-  icon: {fileID: 21300000, guid: bdd4f2fe4ad72894cb4322b83b146008, type: 3}
-  consumable: 0
-  stackable: 0
-  keyItem: 0
-  equipable: 1
-  randomRange: 0
-  maxStack: 0
-  amount: 0
-  amountRange: {x: 0, y: 0}
-  singleUse: 0
-  keyID: 
-  itemID: Flashlight
