--- Assets/Scriptable Objects/Flashlight.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7dd7ddbb99bcdab449a592515f43afb2
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
