--- Assets/Scriptable Objects/Heavy Ammo.asset
@@ -1,29 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2fd88fe6ee8f524418b24e29f14bd1bb, type: 3}
-  m_Name: Heavy Ammo
-  m_EditorClassIdentifier: 
-  _index: 2
-  itemName: Heavy Ammo
-  description: A heavy caliber bullet, it has a great stopping power but is large.
-  icon: {fileID: 21300000, guid: 0bead4418fdba1c4da7030633813256f, type: 3}
-  consumable: 1
-  stackable: 1
-  keyItem: 0
-  equipable: 0
-  randomRange: 1
-  maxStack: 5
-  amount: 1
-  amountRange: {x: 1, y: 2}
-  singleUse: 0
-  keyID: 
-  itemID: HeavyAmmo
