--- Assets/Scriptable Objects/Heavy Ammo.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 7e9fb5e05b47620409a918592dcce9de
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
