+++ Assets/Scriptable Objects/Items.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a007e983d8d201a4792eafc0bea6bb70
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
