+++ Assets/Scriptable Objects/Items/House Key.asset
@@ -1,0 +1,30 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!114 &11400000
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 0}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2fd88fe6ee8f524418b24e29f14bd1bb, type: 3}
+  m_Name: House Key
+  m_EditorClassIdentifier: 
+  _index: 4
+  itemName: House Key
+  description: A extravagent Key made of brass. Maybe someone around here knows where
+    it can be used.
+  icon: {fileID: 21300000, guid: f9e4c141412ca6345963f9ece84aae0a, type: 3}
+  consumable: 0
+  stackable: 0
+  keyItem: 1
+  equipable: 0
+  randomRange: 0
+  maxStack: 0
+  amount: 0
+  amountRange: {x: 0, y: 0}
+  singleUse: 0
+  keyID: House Key
+  itemID:
