+++ Assets/Scriptable Objects/Items/House Key.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c4a76809e236a624fb97be22af0e77da
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
