+++ Assets/Scriptable Objects/Items/Light Ammo.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3a24ac2306ae5404c969ed11ab64d201
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
