+++ Assets/Scriptable Objects/Items/Medium Ammo.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: ea324190f123b6146a548269613a4b5e
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
