+++ Assets/Scriptable Objects/Items/Shotgun.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 65e9f64be63beea498885302e66ab34d
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
