+++ Assets/Scriptable Objects/Items/Small Pistol.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3a59e6cba5366a146ae06856b8d11752
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
