--- Assets/Scriptable Objects/Light Ammo.asset
@@ -1,29 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2fd88fe6ee8f524418b24e29f14bd1bb, type: 3}
-  m_Name: Light Ammo
-  m_EditorClassIdentifier: 
-  _index: 0
-  itemName: Light Ammo
-  description: A small caliber bullet that is versitle and usable in many situations.
-  icon: {fileID: 21300000, guid: a1bc7ae6885093c43af4b74731a7ba65, type: 3}
-  consumable: 1
-  stackable: 1
-  keyItem: 0
-  equipable: 0
-  randomRange: 1
-  maxStack: 15
-  amount: 5
-  amountRange: {x: 4, y: 6}
-  singleUse: 0
-  keyID: 
-  itemID: LightAmmo
