--- Assets/Scriptable Objects/Medium Ammo.asset
@@ -1,30 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2fd88fe6ee8f524418b24e29f14bd1bb, type: 3}
-  m_Name: Medium Ammo
-  m_EditorClassIdentifier: 
-  _index: 1
-  itemName: Medium Ammo
-  description: A medium caliber bullet with general applications. It gets the job
-    done when used and usually puts whatever it's shot at, down.
-  icon: {fileID: 21300000, guid: 33f0b74cefdf97d479ef8c7243bece34, type: 3}
-  consumable: 1
-  stackable: 1
-  keyItem: 0
-  equipable: 0
-  randomRange: 1
-  maxStack: 8
-  amount: 3
-  amountRange: {x: 2, y: 5}
-  singleUse: 0
-  keyID: 
-  itemID: MediumAmmo
