--- Assets/Scriptable Objects/New Shader Values.asset
@@ -1,41 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 1b89d82fa527483428ccf1de21a105c6, type: 3}
-  m_Name: New Shader Values
-  m_EditorClassIdentifier: 
-  spread: 0.5
-  redColorCount: 2
-  greenColorCount: 2
-  blueColorCount: 2
-  bayerLevel: 0
-  ditherDownSamples: 0
-  pointFilterDown: 0
-  exposure: {x: 1, y: 1, z: 1}
-  temperature: 0
-  tint: 0
-  contrast: {x: 1, y: 1, z: 1}
-  linearMidPoint: {x: 0.5, y: 0.5, z: 0.5}
-  brightness: {x: 0, y: 0, z: 0}
-  colorFilter: {r: 0, g: 0, b: 0, a: 0}
-  saturation: {x: 1, y: 1, z: 1}
-  debugMask: 0
-  focalOffset: {x: 0, y: 0}
-  radius: {x: 1, y: 1}
-  hardness: 1
-  intensity: 1
-  channelOffsets: {x: 0, y: 0, z: 0}
-  threshold: 1
-  softThreshold: 0.5
-  bloomDownSamples: 1
-  downSampleDelta: 1
-  upSampleDelta: 0.5
-  bloomIntensity: 1
