--- Assets/Scriptable Objects/Purple Herb.asset
@@ -1,30 +1,0 @@
-%YAML 1.1
-%TAG !u! tag:unity3d.com,2011:
---- !u!114 &11400000
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 0}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 2fd88fe6ee8f524418b24e29f14bd1bb, type: 3}
-  m_Name: Purple Herb
-  m_EditorClassIdentifier: 
-  _index: 3
-  itemName: Purple Herb
-  description: A mysterious purple plant. The numbing sensation you get from it makes
-    you feel better and forget the scrapes and bruises you have.
-  icon: {fileID: 21300000, guid: bc601b3e3ced5b640959fc6b9a6bae5a, type: 3}
-  consumable: 1
-  stackable: 0
-  keyItem: 0
-  equipable: 0
-  randomRange: 0
-  maxStack: 0
-  amount: 0
-  amountRange: {x: 0, y: 0}
-  singleUse: 0
-  keyID: 
-  itemID: Purple Herb
