--- Assets/Scriptable Objects/Purple Herb.asset.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: bd2b58c7f28327140b9c37f6fdbd4c07
-NativeFormatImporter:
-  externalObjects: {}
-  mainObjectFileID: 11400000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
