+++ Assets/Scriptable Objects/Shader.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 39b69105ce3d25b499bd8945c8d7b9df
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
