+++ Assets/Scriptable Objects/Shader/New Shader Values.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 026f267eead465b48b2f2db494e1bd39
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
