--- Assets/Scripts/Player/PlayerAttackHandler.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e195c91668f02ef4db11c3c68cb2d848
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
